requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
yfinance>=0.2.52
pandas>=2.2.2
tabulate>=0.9.0